annbatch: validation densify fix, control_in_memory, loader unify + adapter rename
- fix streaming validation: DAGEvalAdapter densifies eval source/target (with the cupy GPU
  read path annbatch yields a jax sparse CSR that solver.predict can't index).
- add CellFlowAnnbatch.prepare_data(control_in_memory=...): materialize the matched controls
  into a RAM-resident AnnData (a second Scheme source) while the perturbed target streams out
  of core; handles dense arrays and anndata sparse-zarr-group backings. Cuts dataloader ~1/3.
- unify loaders: drop the unused per-split `_annbatch_loaders` dict — train streams via
  `_dataloader`, val/test read via `split_eval_loaders` (DAGEvalLoader).
- rename adapters for symmetry, both now in `_dataloader.py`: DAGLoaderAdapter -> DAGTrainAdapter,
  AnnbatchValidationSampler -> DAGEvalAdapter.
- move TrainingData to `_legacy.py` (in-memory-path container).
