+++ tests/model/test_annbatch_split.py
@@ -87,7 +87,7 @@ class TestPrepareAnnbatchData:
         cf = cellflow.model.CellFlowAnnbatch()
         self._prepare(cf)
         assert cf._scheme is not None and cf._scheme.root == "pert"
-        assert cf._dataloader is not None  # DAGLoaderAdapter wired for train()
+        assert cf._dataloader is not None  # DAGTrainAdapter wired for train()
         assert cf.split_eval_loaders == {}  # no split → no eval loaders
         # condition embeddings assembled (drug is categorical → one-hot), data dim from X
         assert set(cf._condition_data) == {"drug"}
